fix setup process
